# Jpeg encoder optimization via removing second transpose call from FDCT