### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Closes #1778 for the encoder part, decoder part was implemented long time ago so it can be closed after this PR.

Basically, this removes second matrix transpose call from FDCT-Quantize-Huffman pipeline in the jpeg encoder. Performance gain is not **that** big but it's still nice:

Current master
```
// Encoding q=98 | color=YCbCrRatio444
// Elapsed: 11934ms across 1000 iterations
// Average: 11,934ms
```

PR, roughly **4% faster**
```
// Encoding q=98 | color=YCbCrRatio444
// Elapsed: 11457ms across 1000 iterations
// Average: 11,457ms
```

Tests were done using [Calliphora.jpg](https://github.com/SixLabors/ImageSharp/blob/master/tests/Images/Input/Jpg/baseline/Calliphora.jpg) from test suite. Quality 98, no chroma subsampling.

This optimization has 'constant' delta independent of output quality so lower qualities gain better performance boost:

Current master
```
// Encoding q=75 | color=YCbCrRatio444
// Elapsed: 9386ms across 1000 iterations
// Average: 9,386ms
```

PR, roughly **5% faster**
```
// Encoding q=75 | color=YCbCrRatio444
// Elapsed: 8900ms across 1000 iterations
// Average: 8,9ms
```