@br3aker 2022-01-26
@gfoidl sorry for pinging but this may interest you.

AVX implementation has a bit different asm output and performance on `core 3.1` and `.net 6.0`:

**.NET Core 3.1.21** (CoreCLR 4.700.21.51404, CoreFX 4.700.21.51508), X64 RyuJIT
|          Method |     Mean |     Error |    StdDev | Code Size |
|---------------- |---------:|----------:|----------:|----------:|
| &#39;Avx optimized&#39; | 5.666 ns | 0.1440 ns | 0.2018 ns |     478 B |

**.NET Core 6.0.0** (CoreCLR 6.0.21.20104, CoreFX 6.0.21.20104), X64 RyuJIT
|          Method |     Mean |     Error |    StdDev | Code Size |
|---------------- |---------:|----------:|----------:|----------:|
| &#39;Avx optimized&#39; | 7.009 ns | 0.1292 ns | 0.1209 ns |     588 B |

I'm not even close to understand assembly nuances but `.net 6.0` does mask loads a bit different than `core 3.1`. Take a look at first 2 rows calculation:

`core 3.1`
```asm
sub       rsp,58
vzeroupper
vmovaps   [rsp+40],xmm6
vmovaps   [rsp+30],xmm7
vmovaps   [rsp+20],xmm8
vmovaps   [rsp+10],xmm9
xor       eax,eax
mov       [rsp+8],rax
mov       rax,26ECE25CA20
mov       [rsp+8],rax
vmovupd   ymm0,[rcx]
vmovupd   ymm1,[rcx+20]
vmovupd   ymm2,[rcx+40]
vmovupd   ymm3,[rcx+60]
vmovdqu   ymm4,ymmword ptr [rax]
vpermd    ymm5,ymm4,ymm1
vpermd    ymm4,ymm4,ymm0
vmovdqu   ymm6,ymmword ptr [rax+20]
vpshufb   ymm4,ymm4,ymm6
vmovdqu   ymm6,ymmword ptr [rax+40]
vpshufb   ymm5,ymm5,ymm6
vmovdqu   ymm6,ymmword ptr [rax+60]
vpermd    ymm7,ymm6,ymm2
vmovdqu   ymm8,ymmword ptr [rax+80]
vpshufb   ymm8,ymm7,ymm8
vpor      ymm5,ymm5,ymm8
vpor      ymm4,ymm4,ymm5
...
```

`.net 6.0`
```asm
sub       rsp,58
vzeroupper
vmovaps   [rsp+40],xmm6
vmovaps   [rsp+30],xmm7
vmovaps   [rsp+20],xmm8
vmovaps   [rsp+10],xmm9
mov       rax,1CDF57ECA60
mov       [rsp+8],rax
vmovupd   ymm0,[rcx]
vmovupd   ymm1,[rcx+20]
vmovupd   ymm2,[rcx+40]
vmovupd   ymm3,[rcx+60]
vmovdqu   ymm4,ymmword ptr [rax]
vpermd    ymm5,ymm4,ymm1
vpermd    ymm4,ymm4,ymm0
mov       rax,1CDF57ECA80
vmovdqu   ymm6,ymmword ptr [rax]
vpshufb   ymm4,ymm4,ymm6
mov       rax,1CDF57ECAA0
vmovdqu   ymm6,ymmword ptr [rax]
vpshufb   ymm5,ymm5,ymm6
mov       rax,1CDF57ECAC0
vmovdqu   ymm6,ymmword ptr [rax]
vpermd    ymm7,ymm6,ymm2
mov       rax,1CDF57ECAE0
vmovdqu   ymm8,ymmword ptr [rax]
vpshufb   ymm8,ymm7,ymm8
vpor      ymm5,ymm5,ymm8
vpor      ymm4,ymm4,ymm5
...
```

Aren't those multiple `mov rax addr` slower than `[rax + const]`?