@antonfirsov 2022-01-26
The test failure is result of eager merge of #1960. I will revert bbbf687 later.