@gfoidl 2022-01-26
> Aren't those multiple `mov rax addr` slower than using `[rax + const]` directly?

IMO `[rax + const]` should be prefered (as you expect too). Code size is also quite smaller in .NET Core 3.1. (by 110 bytes). So it looks like a regression in .NET 6.
When I look with llvm-mca at the posted assembly, the .NET Core 3.1 version has about 400 µOps less to execute -- this reflects in the shown perf-numbers too.

Can we get a minimal repro so that an issue in dotnet/runtime can be filed?

(Don't hesitate to ping me, but in the next days I'm a bit slow to respond (out of office)).