@br3aker 2022-01-26
> Can we get a minimal repro so that an issue in dotnet/runtime can be filed?

I'll play around with it and ping with results (tomorrow, probably), thanks!