@saucecontrol 2022-01-28
Here's a simple repro for that codegen issue, along with "one weird trick" to work around it 😂

[Sharplab](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8gOgCUBXAOwwEt8YLAMIR8AB14AbGFADKMgG68wMXAG4a9Jq049+ggJI8ZEMfKhKV6zY2btufASyMYovLrmXXatnQ/3OPG4eXiwAGgAcSBrUmgDM2kgM3LjYAGYwTKQMQgwA3jQMRQxibgrYGJnMSWww2AAmAPJckgCesmLYXAA8wK2VAHwMfZW4DAC8Q1wwAO7D/TAA2gC6+YXFG3RoDOTbpNtx2+gMAKzbSNsA7GjrG0VbO3sHR9tnDBcMl7cMAL4xG99iAlqgx3BgGABZAAUAEo1tQ7sURjAAFQlVwTBhQ5EomEAVQ86UEAEFcAAFCBgmRQ2BpSEwfDQVoQ7BQXAAC2wkhYAHEYBhahlYFwVNiFrgYTD/oiGN8NsRLgxZLgYKQWMT6vUoXKZcrVSwADIQBoANRgYAw0HIpAiULErhh6twLjipFhNwRMo2erVRtN5stUGttvtUAYAGoGGaLVabb0FgNhBAHI7SS63TCdUVHQAVCCyMBc1mw6VFH40QHA8hJMEMABysPhXtp80qwzD4wYLYJqQy6s1NJgdIhDKZLLZnO5fIFQ5kMBFMDFoxh2zoUqzss9iIVSpVao1Wo3Gx7RKd8cq22jgeD3TBAwGg7pwCgK6PxRPfdJ55gl4DsYit48Pej4MB+JIDi2z6/jGQZxsiiYiCmmZbncub5oWkjFuuKHltQPxAA===)