@br3aker 2022-01-28
Got a bit ill so couldn't work these 2 days.

> "one weird trick" to work around it

I've actually tried to do entire thing with 'ref pointers' and got worse performance, didn't check assembly though.