@saucecontrol 2022-01-28
The codegen is extremely finicky.  Note the `Unsafe.Add` with `0` offset in that example.  Take that out, and it regresses again.  We're discussing that over in the #lowlevel channel on csharp discord.  You may want to get in on that https://discord.gg/csharp

Hope you feel better!